Straighten out the time logic when using no_std Durations
Since the Reference defaults to starting at Duration::new(0, 0), we
can't move back in time to "before" that point, and so all tests need
to update to use some reference after that.

Also, add a reasonable saturating_sub method to the Reference trait,
remove the Sub requirement and use saturating_sub for all our time
maths in algorithms & states.